move benchamrked kernel to a common file
